refactor(core): split run/entity validation and accepted write phases
